added functions to settings.py to help simplify main.py
