+++ assets/__init__.py
@@ -1,3 +1,2 @@
-from assets.ant_nest import AntNest
+from assets.ant_nest import AntNest, Ant
 from assets.food import Food
-from assets.settings import Settings
